# packages/api-contracts/openapi.json
{
  "components": {
    "schemas": {
      "AccountDeviceBind": {
        "properties": {
          "confirmationNote": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Confirmationnote",
            "type": "string"
          },
          "deviceId": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Deviceid",
            "type": "string"
          }
        },
        "required": [
          "deviceId",
          "confirmationNote"
        ],
        "title": "AccountDeviceBind",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlatformAccountCreate": {
        "properties": {
          "authorizationBasis": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Authorizationbasis",
            "type": "string"
          },
          "displayLabel": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Displaylabel",
            "type": "string"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "externalSubjectRef": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Externalsubjectref",
            "type": "string"
          },
          "platform": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Platform",
            "type": "string"
          },
          "secretRef": {
            "maxLength": 512,
            "minLength": 8,
            "title": "Secretref",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "externalSubjectRef",
          "displayLabel",
          "secretRef",
          "authorizationBasis"
        ],
        "title": "PlatformAccountCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CloudCtl Control API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "operationId": "list_platform_accounts_api_v1_accounts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Platform Accounts Api V1 Accounts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Platform Accounts"
      },
      "post": {
        "operationId": "create_platform_account_api_v1_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformAccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Platform Account Api V1 Accounts Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Platform Account"
      }
    },
    "/api/v1/accounts/{account_id}/bindings": {
      "post": {
        "operationId": "bind_account_device_api_v1_accounts__account_id__bindings_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeviceBind"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bind Account Device Api V1 Accounts  Account Id  Bindings Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bind Account Device"
      }
    },
    "/api/v1/accounts/{account_id}/bindings/{device_id}": {
      "delete": {
        "operationId": "unbind_account_device_api_v1_accounts__account_id__bindings__device_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unbind Account Device"
      }
    },
    "/api/v1/accounts/{account_id}/ownership": {
      "get": {
        "operationId": "get_account_ownership_api_v1_accounts__account_id__ownership_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Account Ownership Api V1 Accounts  Account Id  Ownership Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Account Ownership"
      }
    },
    "/api/v1/accounts/{account_id}:status": {
      "post": {
        "operationId": "update_account_status_api_v1_accounts__account_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Account Status Api V1 Accounts  Account Id  Status Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Account Status"
      }
    },
    "/api/v1/apk-artifacts": {
      "post": {
        "operationId": "register_apk_artifact_api_v1_apk_artifacts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApkArtifactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Apk Artifact Api V1 Apk Artifacts Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Apk Artifact"
      }
    },
    "/api/v1/audit-events": {
      "get": {
        "operationId": "list_audit_events_api_v1_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Audit Events Api V1 Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Events"
      }
    },
    "/api/v1/automation-packages": {
      "post": {
        "operationId": "register_automation_package_api_v1_automation_packages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationPackageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Automation Package Api V1 Automation Packages Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Automation Package"
      }
    },
...
  }
}